There are bugs in champ's delete. This partially fixes them.
Move Champ tests to a test class
